Replace one vs all with one vs one; show max either direction.
Instead of doing a one vs all comparison for each block,
resulting in a lot of repeated comparisons, just
do the unique 1 vs 1 comparisons. This should be faster
for large packages. To handle finding the max, because
I'm too stupid to figure out exactly how you find the max
pairwise based only on the two indices, I filter to find
the max a->b and max b->a similarity. This will produce
a slightly longer list but hopefully should still be
informative because I am also removing the
duplication in output lists
(ie, Block A is most similar to Block B, and Block B is most
similar to Block B).

There may be a smarter way to do this. In fact, it might be
worthwhile keeping all combns and just filtering by
similarity score above x (eg, above 0.6, or above some null
distn calculated based on block size).