ensure no replicated block-pairs and no cycles
When searching for duplicated code blocks, for each block, dupree
(currently) returns the most-similar other code-block. This led to
double reporting of some code-block pairs. Alan O'Callaghan
provided an initial fix to prevent double reporting.

However, that solution led to some cycles in the code-block graph
A test was added to catch a simple example of such cycles. And a
fix, based on the original one-against-all use of seq_sim was
added.

The results presented in README.md have changed