Replace one vs all with one vs one; show max either direction. (#30)
* Replace one vs all with one vs one; show max either direction.

Instead of doing a one vs all comparison for each block,
resulting in a lot of repeated comparisons, just
do the unique 1 vs 1 comparisons. This should be faster
for large packages. To handle finding the max, because
I'm too stupid to figure out exactly how you find the max
pairwise based only on the two indices, I filter to find
the max a->b and max b->a similarity. This will produce
a slightly longer list but hopefully should still be
informative because I am also removing the
duplication in output lists
(ie, Block A is most similar to Block B, and Block B is most
similar to Block B).

There may be a smarter way to do this. In fact, it might be
worthwhile keeping all combns and just filtering by
similarity score above x (eg, above 0.6, or above some null
distn calculated based on block size).

* add prescribed ordering for reciprocal-best-match

Added tests to ensure that when two code blocks are
reciprocal-best-matches only one row of data for these blocks is
present in the result data-frame returned by dupree.

When this happens, we ensure that file_a <= file_b and (if
file_a == file_b) that block_a <= block_b.

Modified the initialisation of `EnumeratedCodeTable` to ensure that
the input code blocks are ordered by file-name and then block
number.

Also styled some code, ensured new depencies were imported and
remade the README.md (although README.Rmd has not changed, the
results tables presented in the README file have changed)

* add test: when returning the best match for each input block, there should be no cycles in the results

* ensure no replicated block-pairs and no cycles

When searching for duplicated code blocks, for each block, dupree
(currently) returns the most-similar other code-block. This led to
double reporting of some code-block pairs. Alan O'Callaghan
provided an initial fix to prevent double reporting.

However, that solution led to some cycles in the code-block graph
A test was added to catch a simple example of such cycles. And a
fix, based on the original one-against-all use of seq_sim was
added.

The results presented in README.md have changed

* bump dev-version and added NEWS

* fix linting