fix tests that call to `seq_sim`
In a recent commit on stringdist, the author (@markvanderloo) found that
lazily-evaluated arguments to `seq_dist` in testthat/tinytest tests
caused those tests to fail.

Here we were receiving intermittent 'Warning: stack imbalance ...'
warnings in tests over code that calls to `stringdist::seq_sim`.

To get these tests to pass repeatedly, we replaced all instances of 1:3
or seq(3) (or similar) with explicitly constructed vectors like
c(1,2,3).

Although this appears to allow the tests to run, I have no real
understanding of why the modification works and am worried that there
may be an underlying problem that may cause bigger problems down the
line.